+++ samples/trade360-samples/src/main/java/eu/lsports/trade360feedexample/RabbitMQFeed.java
@@ -59,7 +59,7 @@ public class RabbitMQFeed {
     // - The association between Inplay Rabbit connection factory and this method is made by bean name written in containerFactory annotation properties. Name defined in application properties
    // - Name of queue is taken fom application properties
     // - Error handler is set by errorHandler annotation properties
-    @RabbitListener(containerFactory = "${rabbitmq.inplay.rabbit_listener_container_factory_name}", queues = "#{@inPlayRabbitConnectionConfiguration.consumeQueueName}", errorHandler="inplayErrorMessageHandler")
+    @RabbitListener(containerFactory = "${rabbitmq.inplay.rabbit_listener_container_factory_name}",  queues = "_${rabbitmq.inplay.package_id}_", errorHandler="inplayErrorMessageHandler")
     public void inPlayProcessMessage(final Message amqpMessage, Channel channel, @Header(AmqpHeaders.DELIVERY_TAG) long tag) throws Exception {
         inPlayMessageHandler.process(amqpMessage);
 
@@ -72,7 +72,7 @@ public class RabbitMQFeed {
     // - The association between Prematch Rabbit connection factory and this method is made by bean name written in containerFactory annotation properties. Name defined in application properties
     // - Name of queue is taken fom application properties
     // - Error handler is set by errorHandler annotation properties
-    @RabbitListener(containerFactory = "${rabbitmq.prematch.rabbit_listener_container_factory_name}", queues = "#{@preMatchRabbitConnectionConfiguration.consumeQueueName}", errorHandler="prematchErrorMessageHandler")
+    @RabbitListener(containerFactory = "${rabbitmq.prematch.rabbit_listener_container_factory_name}", queues = "_${rabbitmq.prematch.package_id}_", errorHandler="prematchErrorMessageHandler")
     public void preMatchProcessMessage(final Message message, Channel channel,
                                        @Header(AmqpHeaders.DELIVERY_TAG) long tag) throws Exception {
         preMatchMessageHandler.process(message);