+++ sdk/trade360-java-sdk/src/main/java/eu/lsports/trade360_java_sdk/feed/rabbitmq/configurations/RabbitMqConnectionConfigurationValidator.java
@@ -10,7 +10,6 @@ public final class RabbitMqConnectionConfigurationValidator {
 
     static void validate(RabbitConnectionConfiguration configuration) {
         validateQueueConfiguration(configuration);
-        RabbitMqSslConfigurationValidator.validate(configuration);
     }
 
     private static void validateQueueConfiguration(RabbitConnectionConfiguration configuration) {